requirements.txt
flask
ultralytics
numpy
Pillow
